Move model to model file
