+++ app/models.py
@@ -1,3 +1,4 @@
+from pydantic import BaseModel
 from sqlalchemy import Column, Integer, Float, String
 from .database import Base
 
@@ -11,3 +12,16 @@ class TaxInfo(Base):
     tax_amount = Column(Float, nullable=False)
     tax_rate = Column(Float, nullable=False)
     description = Column(String, nullable=True)
+
+
+class TaxInfoResponse(BaseModel):
+    id: int
+    income: float
+    expenses: float
+    tax_amount: float
+    tax_rate: float
+    description: str | None = None
+
+    class Config:
+        orm_mode = True
+        from_attributes = True
